# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: Evernest API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Baby tracking backend. All requests/responses are JSON. All authenticated
    endpoints require `Authorization: Bearer <access_token>`. The refresh
    token is returned in the JSON body of `/v1/auth/{register,login,refresh}`
    and the client persists it itself (e.g. `localStorage`); it is passed
    back in the body of `/v1/auth/{refresh,logout}`. The legacy
    `evernest_refresh` `httpOnly` cookie is still set and still accepted as
    a fallback during a migration window — see the `RefreshRequest` and
    `refreshCookie` schemas — but new clients should ignore the cookie.

    ## Idempotency

    Every event POST (bottle-feeds, diapers, pumpings, nursing-sessions,
    growths, sleeps) accepts a client-supplied `id` (UUIDv7) in the request body.
    Inserts use `ON CONFLICT (id) DO NOTHING`, so retrying a write with the
    same `id` returns the existing row with status `201` instead of
    erroring. If `id` is omitted the server generates one. There is no
    separate `Idempotency-Key` header — the body field IS the idempotency
    key.

    ## Errors

    All non-2xx responses share the same envelope:

    ```json
    { "error": { "code": "validation_failed", "message": "..." } }
    ```

    `code` is a stable machine-readable string (e.g. `validation_failed`,
    `unauthorized`, `not_found`). `message` is human-readable.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get the authenticated user's profile
      responses:
        "200":
          description: The current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/households:
    post:
      tags: [Households]
      operationId: createHousehold
      summary: Create a household
      description: |
        Creates a household and adds the caller as its `owner` member.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHouseholdRequest"
      responses:
        "201":
          description: Household created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Households]
      operationId: listHouseholds
      summary: List households the caller is a member of
      responses:
        "200":
          description: Households (oldest first).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Household"
              example:
                - id: 01919b86-3a4d-7c8a-9f1d-2c4e6a7b8d9f
                  name: The Arsjad Household
                  role: owner
                  created_at: "2026-04-01T12:34:56Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/invites/{token}:
    parameters:
      - $ref: "#/components/parameters/InviteTokenPath"
    get:
      tags: [Invites]
      operationId: getInviteInfo
      summary: Public lookup of an invite's display metadata
      description: |
        Unauthenticated lookup. Returns only the minimum metadata the FE
        needs to render a "Join {household} as {role}?" confirmation:
        household name, role, and expiry. Member list and inviter
        identity are deliberately omitted. Used / expired / unknown
        tokens uniformly return 404 to avoid token-enumeration leaks.
      security: []
      responses:
        "200":
          description: Invite is pending; metadata returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteInfo"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invites]
      operationId: revokeInvite
      summary: Revoke an invite (owner or creator only)
      description: |
        Deletes the invite, making subsequent accept attempts return 404.
        Authorization: caller must be the inviting household's `owner`
        OR the user who created the invite. Non-authorized callers see
        the same 404 as a missing token to avoid disclosure.
      responses:
        "204":
          description: Revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/babies/{babyID}:
    parameters:
      - $ref: "#/components/parameters/BabyIDPath"
    get:
      tags: [Babies]
      operationId: getBaby
      summary: Get a baby by id
      responses:
        "200":
          description: The baby.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baby"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/bottle-feeds/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    patch:
      tags: [Bottle Feeds]
      operationId: updateBottleFeed
      summary: Update a bottle feed
      description: |
        Partial update. Fields omitted from the request body are left
        unchanged. Server preserves `id`, `source`, `created_at`, and
        `created_by_user_id` regardless of input. To clear `notes`, send
        an empty string; omitting the field leaves notes unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBottleFeedRequest"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BottleFeed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Bottle Feeds]
      operationId: deleteBottleFeed
      summary: Delete a bottle feed
      description: |
        Idempotent: deletes return `204` whether or not the row existed.
      responses:
        "204":
          description: Deleted (or no row matched).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/diapers/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    patch:
      tags: [Diapers]
      operationId: updateDiaper
      summary: Update a diaper change
      description: |
        Partial update. Same contract as `PATCH /v1/bottle-feeds/{id}`. The
        `photo` field follows the same "send empty string to clear" rule as
        `notes`: omit to leave it alone, `""` to clear, or send a fresh
        base64 payload + `photo_mime` to replace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDiaperRequest"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaper"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          description: Photo exceeds the size limit (2 MB raw / 3 MB body).
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Diapers]
      operationId: deleteDiaper
      summary: Delete a diaper change
      description: Idempotent.
      responses:
        "204":
          description: Deleted (or no row matched).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/nursing-sessions/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    patch:
      tags: [Nursing Sessions]
      operationId: updateNursingSession
      summary: Close an open session OR edit a closed one
      description: |
        Dispatches based on the row's current `ended_at` state:

        - **Open session** (`ended_at IS NULL`): closes the session.
          Body must include `ended_at`, `left_duration_s`, and
          `right_duration_s` (all required). This is the legacy
          close-session contract. Returns `409 already_closed` if the
          row was closed by a concurrent request.
        - **Closed session** (`ended_at IS NOT NULL`): partial edit.
          All fields optional. Sending `ended_at: null` is not
          supported (edits can't re-open a closed session). The notes
          and `starting_breast` clear conventions match the other
          event PATCHes. Returns `409 session_state_changed` if the
          row was deleted or re-opened mid-edit.

        Server preserves `id`, `source`, `created_at`, and
        `created_by_user_id` regardless of input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/EndNursingRequest"
                - $ref: "#/components/schemas/UpdateNursingRequest"
      responses:
        "200":
          description: Closed (open session path) or updated (closed session path).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nursing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Either `already_closed` (close-path) or
            `session_state_changed` (edit-path) — see description.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Nursing Sessions]
      operationId: deleteNursingSession
      summary: Delete a nursing session
      description: Idempotent.
      responses:
        "204":
          description: Deleted (or no row matched).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/growths/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    patch:
      tags: [Growths]
      operationId: updateGrowth
      summary: Update a growth measurement
      description: |
        Partial update. Each measurement (`weight_g`, `height_cm`,
        `head_circumference_cm`) can be cleared via its companion
        `clear_*` boolean. The DB's "at least one non-NULL" CHECK is
        mirrored in Go and surfaces as `422` if a clear would empty
        the row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGrowthRequest"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Growth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Growths]
      operationId: deleteGrowth
      summary: Delete a growth measurement
      description: Idempotent.
      responses:
        "204":
          description: Deleted (or no row matched).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/sleeps/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    patch:
      tags: [Sleeps]
      operationId: updateSleep
      summary: Close an open sleep OR edit a closed one
      description: |
        Dispatches based on the row's current `ended_at` state:

        - **Open session** (`ended_at IS NULL`): closes the session.
          Body must include `ended_at` (`>= started_at`). Returns
          `409 already_closed` if the row was closed by a concurrent
          request.
        - **Closed session** (`ended_at IS NOT NULL`): partial edit.
          All fields optional. Sending `ended_at: null` is not
          supported (edits can't re-open a closed session). Set
          `clear_sleep_type: true` to clear the type; send an empty
          string for `location` or `notes` to clear them. Returns
          `409 session_state_changed` if the row was deleted or
          re-opened mid-edit.

        Server preserves `id`, `source`, `created_at`, and
        `created_by_user_id` regardless of input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/EndSleepRequest"
                - $ref: "#/components/schemas/UpdateSleepRequest"
      responses:
        "200":
          description: Closed (open session path) or updated (closed session path).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sleep"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Either `already_closed` (close-path) or
            `session_state_changed` (edit-path) — see description.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Sleeps]
      operationId: deleteSleep
      summary: Delete a sleep session
      description: Idempotent. Open rows can be deleted ("started by mistake").
      responses:
        "204":
          description: Deleted (or no row matched).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # -----------------------------------------------------------------------
  # Coming soon
  # -----------------------------------------------------------------------
  # CP3 Charts will add `GET /v1/babies/{babyID}/charts/daily` in its own
  # PR. The schema for that endpoint is intentionally NOT defined here —
  # see the parallel CP3 Charts agent's PR for the full spec.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HMAC-signed JWT access token issued by `/v1/auth/{register,login,refresh}`.
        Send as `Authorization: Bearer <token>`.
    refreshCookie:
      type: apiKey
      in: cookie
      name: evernest_refresh
      description: |
        DEPRECATED — accepted as a fallback during a migration window so
        clients that pre-date the body-based refresh-token rollout keep
        working. New clients should ignore this cookie and use the
        `refresh_token` field in `TokenResponse` + `RefreshRequest`
        instead.

        `httpOnly`, opaque (random) refresh token. Set by
        `/v1/auth/{register,login,refresh}` and consumed (as a fallback)
        by `/v1/auth/{refresh,logout}`. Cookie path is `/v1/auth`.
...
  parameters:
...
    InviteTokenPath:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 64
        pattern: ^[A-Za-z0-9_-]+$
      example: 3xK2-9aBcD8fE7gH6iJ5kL
      description: |
        URL-safe base64 invite token (`crypto/rand` 16 bytes → 22 chars,
        no padding). Treat it as a one-time secret; only the SHA-256 hash
        is persisted server-side.
    BabyIDPath:
      name: babyID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 01919b87-9a8c-7f2d-b1c3-4d5e6f7a8b9c
      description: Baby UUID.
    IDPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 01919b88-1c2d-7e3f-a4b5-c6d7e8f9a0b1
      description: Resource UUID (event id).
...
  schemas:
...
    User:
      type: object
      required: [id, email, display_name, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 01919b85-1c2d-7e3f-a4b5-c6d7e8f9a0b1
        email:
          type: string
          format: email
          example: victor@example.com
        display_name:
          type: string
          example: Victor
        created_at:
          type: string
          format: date-time
          example: "2026-04-01T12:34:56Z"
      example:
        id: 01919b85-1c2d-7e3f-a4b5-c6d7e8f9a0b1
        email: victor@example.com
        display_name: Victor
        created_at: "2026-04-01T12:34:56Z"
...
    Household:
      type: object
      required: [id, name, role, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 01919b86-3a4d-7c8a-9f1d-2c4e6a7b8d9f
        name:
          type: string
          example: The Arsjad Household
        role:
          type: string
          description: |
            The caller's role in this household. `owner` can create + revoke
            invites and is granted on household creation; `caregiver` can
            see all data and log events but cannot invite new members.
          enum: [owner, caregiver]
          example: owner
        created_at:
          type: string
          format: date-time
          example: "2026-04-01T12:34:56Z"
    CreateHouseholdRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          example: The Arsjad Household
...
    InviteInfo:
      type: object
      required: [household_name, role, expires_at]
      properties:
        household_name:
          type: string
          example: The Arsjad Household
        role:
          type: string
          enum: [owner, caregiver]
          example: caregiver
        expires_at:
          type: string
          format: date-time
...
    Baby:
      type: object
      required: [id, household_id, name, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 01919b87-9a8c-7f2d-b1c3-4d5e6f7a8b9c
        household_id:
          type: string
          format: uuid
          example: 01919b86-3a4d-7c8a-9f1d-2c4e6a7b8d9f
        name:
          type: string
          example: Mira
        date_of_birth:
          type: string
          format: date
          description: Date in `YYYY-MM-DD`. Omitted if unknown.
          example: "2026-03-15"
        sex:
          type: string
          enum: [female, male, unspecified]
          description: Omitted if unspecified by client (the column is nullable).
          example: female
        created_at:
          type: string
          format: date-time
          example: "2026-04-02T08:00:00Z"
...
    BottleFeed:
      type: object
      required: [id, baby_id, occurred_at, milk_source, amount_ml, source, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 01919b88-1c2d-7e3f-a4b5-c6d7e8f9a0b1
        baby_id:
          type: string
          format: uuid
          example: 01919b87-9a8c-7f2d-b1c3-4d5e6f7a8b9c
        occurred_at:
          type: string
          format: date-time
          example: "2026-05-26T08:30:00Z"
        milk_source:
          type: string
          enum: [breast, formula]
          example: breast
        amount_ml:
          type: number
          format: double
          exclusiveMinimum: 0
          maximum: 2000
          example: 120
        notes:
          type: string
          example: "took it eagerly"
        source:
          type: string
          description: How the row was created. Always `manual` from this endpoint; importers may set other values.
          example: manual
        created_at:
          type: string
          format: date-time
          example: "2026-05-26T08:30:01Z"
...
    UpdateBottleFeedRequest:
      type: object
      description: |
        All fields optional. Omitted fields are left unchanged. Send an
        empty string for `notes` to clear it.
      properties:
        occurred_at:
          type: string
          format: date-time
          example: "2026-05-26T08:30:00Z"
        milk_source:
          type: string
          enum: [breast, formula]
          example: breast
        amount_ml:
          type: number
          format: double
          exclusiveMinimum: 0
          maximum: 2000
          example: 120
        notes:
          type: string
          example: "took it eagerly"
    Diaper:
      type: object
      required: [id, baby_id, occurred_at, type, has_photo, source, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 01919b89-2d3e-7f4a-b5c6-d7e8f9a0b1c2
        baby_id:
          type: string
          format: uuid
          example: 01919b87-9a8c-7f2d-b1c3-4d5e6f7a8b9c
        occurred_at:
          type: string
          format: date-time
          example: "2026-05-26T09:15:00Z"
        type:
          type: string
          enum: [wet, soiled, mixed]
          example: mixed
        notes:
          type: string
          example: "blowout"
        has_photo:
          type: boolean
          description: |
            True when the row has an attached photo. The bytes are fetched
            on demand via `GET /v1/diapers/{id}/photo` — the list endpoint
            never inlines the blob.
          example: false
        photo_mime:
          type: string
          enum: [image/jpeg, image/png, image/webp]
          description: MIME of the attached photo. Omitted when `has_photo` is false.
          example: image/jpeg
        source:
          type: string
          example: manual
        created_at:
          type: string
          format: date-time
          example: "2026-05-26T09:15:01Z"
...
    UpdateDiaperRequest:
      type: object
      description: |
        All fields optional. Omitted fields are left unchanged. Send an
        empty string for `notes` or `photo` to clear it; send fresh
        `photo` + `photo_mime` to replace.
      properties:
        occurred_at:
          type: string
          format: date-time
          example: "2026-05-26T09:15:00Z"
        type:
          type: string
          enum: [wet, soiled, mixed]
          example: mixed
        notes:
          type: string
          example: "blowout"
        photo:
          type: string
          format: byte
          description: |
            Base64-encoded image bytes. Empty string (`""`) clears the
            stored photo. Same 2 MB raw / 3 MB body caps as the create
            endpoint.
        photo_mime:
          type: string
          enum: [image/jpeg, image/png, image/webp]
          description: Required when `photo` is set to a non-empty string.
...
    EndNursingRequest:
      type: object
      required: [ended_at, left_duration_s, right_duration_s]
      description: |
        Body shape for closing an OPEN nursing session via
        `PATCH /v1/nursing-sessions/{id}`. All three fields required.
      properties:
        ended_at:
          type: string
          format: date-time
          example: "2026-05-26T11:25:00Z"
        left_duration_s:
          type: integer
          minimum: 0
          maximum: 21600
          example: 600
        right_duration_s:
          type: integer
          minimum: 0
          maximum: 21600
          example: 900
...
    EndSleepRequest:
      type: object
      required: [ended_at]
      description: |
        Body shape for closing an OPEN sleep session via
        `PATCH /v1/sleeps/{id}`.
      properties:
        ended_at:
          type: string
          format: date-time
          example: "2026-05-26T14:25:00Z"
...
  responses:
    BadRequest:
      description: Malformed request body, query parameter, or path id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            badJSON:
              summary: Body could not be decoded
              value:
                error:
                  code: bad_request
                  message: "json: unknown field \"foo\""
            badID:
              summary: Path id is not a valid UUID
              value:
                error:
                  code: bad_id
                  message: invalid baby id
            badRange:
              summary: from/to query parameter is malformed or inverted
              value:
                error:
                  code: bad_range
                  message: from must be RFC3339
    Unauthorized:
      description: Missing/invalid bearer token (or, on `/v1/auth/refresh`, a missing/invalid refresh cookie).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: missing bearer token
    Forbidden:
      description: The caller is authenticated but not a member of the household that owns the resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: forbidden
              message: not a member of this household
    NotFound:
      description: The household or baby referenced in the path does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: baby not found
    ValidationError:
      description: |
        Request was syntactically valid JSON but failed semantic validation
        (struct tag rules, enum constraints, range bounds, etc.). The
        `message` field is the raw validator output.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingField:
              summary: Required field missing
              value:
                error:
                  code: validation_failed
                  message: "Key: 'createReq.OccurredAt' Error:Field validation for 'OccurredAt' failed on the 'required' tag"
            outOfRange:
              summary: Numeric value out of range
              value:
                error:
                  code: validation_failed
                  message: "Key: 'createReq.AmountML' Error:Field validation for 'AmountML' failed on the 'gt' tag"
            growthAllNull:
              summary: Growth POST with no measurements
              value:
                error:
                  code: validation_failed
                  message: "at least one of weight_g, height_cm, head_circumference_cm is required"
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal
              message: could not create baby